chore: Update imports in paasta_tools to use AutoscalingParamsDict and MetricsProviderDict
